AS7-2721,AS7-2720
relocating test configurations into testsuite/integration/basic
