+++ testsuite/integration/basic/src/test/resources/test-configs/domain-ds.xml
