+++ testsuite/integration/basic/src/test/resources/test-configs/standalone-ds-prop.xml
